requirements.txt
# Backend
fastapi
uvicorn[standard]
python-dotenv
sentence-transformers
transformers
torch
accelerate
pdfplumber
python-docx
pinecone

# Frontend
streamlit
requests